Replaced legacy test, written a new one
